streamlit-app/requirements.txt
streamlit==1.37.0
openai==1.40.0
httpx>=0.25.0,<0.28
azure-identity==1.17.0
azure-monitor-query==1.4.0
python-dotenv==1.0.1
requests==2.32.3
